--- a/src/lib.rs
+++ b/src/lib.rs
@@ -294,7 +294,7 @@ struct Output {
 
 impl Output {
     #[inline(always)]
-    pub fn new(value: u32, length: u32, parent: u32) -> Self {
+    pub const fn new(value: u32, length: u32, parent: u32) -> Self {
         Self {
             value,
             length,