This PR improved the data structure for output sets, which stores output sets on a forest representation.

The new forest representation reduces the memory usage on UniDic, as follows.

```
# old version
== data/words_100000 ==
daachorse (bytewise): 8700816 bytes, 8.298 MiB
daachorse (charwise): 11210880 bytes, 10.692 MiB
== data/unidic/unidic ==
daachorse (bytewise): 36382056 bytes, 34.697 MiB
daachorse (charwise): 30220136 bytes, 28.820 MiB

# new version
== data/words_100000 ==
daachorse (bytewise): 8999820 bytes, 8.583 MiB
daachorse (charwise): 11509884 bytes, 10.977 MiB
== data/unidic/unidic ==
daachorse (bytewise): 29523252 bytes, 28.156 MiB
daachorse (charwise): 23361332 bytes, 22.279 MiB
```

There is no significant difference in matching time.

```
unidic/find_overlapping/daachorse                                                                           
                        time:   [11.510 ms 11.541 ms 11.596 ms]
                        change: [-0.7825% -0.4400% +0.0661%] (p = 0.04 < 0.05)
                        Change within noise threshold.
Found 1 outliers among 30 measurements (3.33%)
  1 (3.33%) high severe
unidic/find_overlapping/daachorse/charwise                                                                            
                        time:   [6.5334 ms 6.5553 ms 6.5945 ms]
                        change: [-2.5997% -2.2480% -1.6692%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 1 outliers among 30 measurements (3.33%)
  1 (3.33%) high severe

words_100000/find_overlapping/daachorse                                                                            
                        time:   [4.3369 ms 4.3394 ms 4.3422 ms]
                        change: [+2.3980% +2.4814% +2.5662%] (p = 0.00 < 0.05)
                        Performance has regressed.
Found 1 outliers among 30 measurements (3.33%)
  1 (3.33%) high mild
words_100000/find_overlapping/daachorse/charwise                                                                            
                        time:   [4.7484 ms 4.7626 ms 4.7781 ms]
                        change: [+3.1210% +3.5562% +3.9811%] (p = 0.00 < 0.05)
                        Performance has regressed.
Found 1 outliers among 30 measurements (3.33%)
  1 (3.33%) high mild
```